+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: pscott <pscott@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 11:41:39 by pscott            #+#    #+#             */
-/*   Updated: 2018/11/07 16:01:32 by pscott           ###   ########.fr       */
+/*   Updated: 2018/11/07 16:10:16 by pscott           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (len - i)
